feat: store whether the page is managed by Mergify in a cache
Create a cache with the localStorage to store whether Mergify manages the page. This cache is updated whenever we detect a Mergify component on the page. It expires after 1 day.

This will resolve a user issue. Mergify isn't detected until the user opens the check-run drawer because GitHub doesn't load the check-runs if the drawer isn't opened. The cache will be updated whenever the user looks for the check-runs on the discussion page. It should improve the user experience.

Fixes MRGFY-4917

Change-Id: I010fbd3386ea3ed3e9577f4f9998cda9fb399abd