feat: add safari signing (#73)
This rework the safari artefact to sign it with the CI.

The signature for now is a Developer ID one, that allows distribution of
the package outside app store.

The process to publish it to the Apple App Store is a bit to tedious for
us to support for now.

Fixes: MRGFY-4982